revert: restore authors page fetching
