+++ src/hooks/useAuthors.ts
@@ -68,18 +68,14 @@ export const useAuthors = () => {
       console.log('Fetching authors with books from database...');
 
       try {
-        const res = await fetch('/api/authors?page=1&pageSize=100', {
-          credentials: 'include',
-        });
+        // Use the new function that only returns authors with books
+        const { data, error } = await supabase.rpc('get_authors_with_books');
 
-        if (!res.ok) {
-          const err: any = new Error('Failed to fetch authors');
-          err.status = res.status;
-          throw err;
+        if (error) {
+          console.error('Failed to fetch authors with books:', error.message || error);
+          console.info('Using fallback author data due to API error');
+          return fallbackAuthors;
         }
-
-        const result = await res.json();
-        const data = result.authors as any[];
 
         if (!data || data.length === 0) {
           console.warn('No authors with books found, using fallback data');
@@ -133,18 +129,24 @@ export const usePaginatedAuthors = (page = 1, pageSize = 10) => {
   return useQuery({
     queryKey: ['paginated-authors', page, pageSize],
     queryFn: async (): Promise<PaginatedAuthors> => {
+      const startIndex = (page - 1) * pageSize;
+      const endIndex = page * pageSize - 1;
+
       try {
-        const res = await fetch(`/api/authors?page=${page}&pageSize=${pageSize}`, {
-          credentials: 'include',
-        });
-        if (!res.ok) {
-          const err: any = new Error('Failed to fetch authors');
-          err.status = res.status;
-          throw err;
+        const { data, error, count } = await supabase
+          .rpc('get_authors_with_books', {}, { count: 'exact' })
+          .range(startIndex, endIndex)
+          .order('name', { ascending: true });
+
+        if (error || !data) {
+          console.error('Failed to fetch paginated authors:', error?.message || error);
+          return {
+            authors: fallbackAuthors.slice(startIndex, endIndex + 1),
+            total: fallbackAuthors.length,
+          };
         }
 
-        const result = await res.json();
-        const authors = (result.authors || []).map((author: any) => ({
+        const authors = data.map((author: any) => ({
           id: author.id,
           name: author.name,
           bio: author.bio || 'Bio will be updated by the author.',
@@ -153,7 +155,7 @@ export const usePaginatedAuthors = (page = 1, pageSize = 10) => {
           website_url: author.website_url,
           social_links: author.social_links || {},
           genres: author.genres || [],
-          books_count: parseInt(author.books_count) || 0,
+          books_count: parseInt(author.actual_books_count) || 0,
           followers_count: author.followers_count || 0,
           rating: parseFloat(author.rating) || 4.0,
           upcoming_events: author.upcoming_events || 0,
@@ -162,16 +164,16 @@ export const usePaginatedAuthors = (page = 1, pageSize = 10) => {
           verified: author.verified || false,
           verification_type: author.verification_type,
           availableSlots:
-            author.books_count > 0
-              ? [`Meet the author of ${author.books_count} book${author.books_count > 1 ? 's' : ''}`]
+            author.actual_books_count > 0
+              ? [`Meet the author of ${author.actual_books_count} book${author.actual_books_count > 1 ? 's' : ''}`]
               : ['Available for consultation'],
         })) as Author[];
 
-        return { authors, total: result.total ?? authors.length };
+        return { authors, total: count ?? authors.length };
       } catch (err) {
         console.error('Exception while fetching paginated authors:', err);
         return {
-          authors: fallbackAuthors.slice(0, pageSize),
+          authors: fallbackAuthors.slice(startIndex, endIndex + 1),
           total: fallbackAuthors.length,
         };
       }
