+++ src/pages/authors/[slug].tsx
@@ -29,12 +29,7 @@ const AuthorSlugPage = () => {
   const { slug } = useParams<{ slug: string }>();
   const { user } = useAuth();
   const { data: author, isLoading } = useAuthorBySlug(slug);
-  const {
-    data: authorBooks,
-    isLoading: booksLoading,
-    error: booksError,
-    refetch: refetchBooks,
-  } = useAuthorBooks(author?.id || '');
+  const { data: authorBooks, isLoading: booksLoading } = useAuthorBooks(author?.id || '');
   const { data: authorPosts, isLoading: postsLoading } = useAuthorPosts(author?.id);
   const { data: questions, isLoading: questionsLoading } = useAuthorQuestions(author?.id);
   const { data: events, isLoading: eventsLoading } = useAuthorEvents(author?.id);
@@ -181,7 +176,7 @@ const AuthorSlugPage = () => {
                 <CardContent>
                   <div className="prose prose-gray dark:prose-invert max-w-none">
                     <p className="text-muted-foreground leading-relaxed text-base">
-                      {author.bio ? author.bio : 'Bio not yet provided'}
+                      {author.bio ? author.bio : "This author hasn\u2019t added a bio yet. Check back soon!"}
                     </p>
                   </div>
                   
@@ -191,7 +186,7 @@ const AuthorSlugPage = () => {
                         <div className="font-medium text-foreground">Location</div>
                         <div className="text-muted-foreground flex items-center gap-1">
                           <MapPin className="w-3 h-3" />
-                          {author.location || 'Location not specified'}
+                          {author.location}
                         </div>
                       </div>
                       <div>
@@ -307,11 +302,6 @@ const AuthorSlugPage = () => {
                           </CardContent>
                         </Card>
                       ))}
-                    </div>
-                  ) : booksError ? (
-                    <div className="text-center py-12 space-y-4">
-                      <p className="text-muted-foreground">Couldn't load books. Please try again.</p>
-                      <Button onClick={() => refetchBooks()}>Retry</Button>
                     </div>
                   ) : (
                     <div className="text-center py-12">
